requirements.txt
pygame-ce>=2.5.0
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
python-chess>=1.999
websocket-client>=1.6.0
